requirements.txt
python-can>=4.0.0
RPi.GPIO>=0.7.0